+++ src/logger.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+
+export const logger = {
+  log(...args) {
+    console.log(...args);
+  },
+  error(...args) {
+    console.error(...args);
+  },
+};